--- docs/api/pylabrobot.btx.rst
@@ -1,45 +1,0 @@
-.. currentmodule:: pylabrobot.btx
-
-pylabrobot.btx package
-======================
-
-Gemini X2
----------
-
-.. currentmodule:: pylabrobot.btx.gemini_x2
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    BTXGeminiX2
-    BTXGeminiX2Driver
-    BTXGeminiX2ElectroporationBackend
-
-.. currentmodule:: pylabrobot.btx.file_transfer_control
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    FileTransferControl
-
-.. currentmodule:: pylabrobot.btx.the_ghost_touch
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    TheGhostTouch
-
-.. currentmodule:: pylabrobot.btx.ht200
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    BTXHT200
