+++ docs/api/pylabrobot.rst
@@ -24,7 +24,6 @@ Manufacturers
     pylabrobot.azenta
     pylabrobot.big_bear
     pylabrobot.brooks
-    pylabrobot.btx
     pylabrobot.byonoy
     pylabrobot.cole_parmer
     pylabrobot.curiox
