+++ docs/user_guide/index.md
@@ -32,7 +32,6 @@ agilent/index
 azenta/index
 big_bear/index
 brooks/index
-btx/index
 byonoy/index
 cole_parmer/index
 curiox/index
