+++ docs/user_guide/thermo_fisher/btx/index.md
@@ -3,7 +3,8 @@
 ```{toctree}
 :maxdepth: 1
 
-gemini_x2/hello-world
+gemini/X2/installation
+gemini/X2/hello-world
 ```
 
 ## Gemini X2
@@ -11,4 +12,3 @@ gemini_x2/hello-world
 | Model | PLR device | Capabilities |
 |---|---|---|
 | Gemini X2 | `BTXGeminiX2` | Electroporation |
-
