+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/01_user_protocols_top.png
