+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/02_protocol_summary.png
