+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/03_run_protocol_prerun.png
