+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/04_set_plate_columns_open.png
