+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/05_set_plate_columns_after_first_confirm.png
