+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/06_set_plate_columns_confirmed_run_view.png
