+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/07_go_prerun.png
