+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/08_go_delivering_pulse.png
