+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/09_go_pulses_delivered.png
