+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/10_returned_home_after_go.png
