+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/metadata.json
