+++ pylabrobot/thermo_fisher/btx/gemini/X2/test_data/gemini_x2/screens/user_protocols_double_up_inactive.png
